requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
environs==8.0.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
jiter==0.8.2
loader==2017.9.11
magic-filter==1.0.12
marshmallow==3.20.2
multidict==6.0.4
openai==1.59.7
packaging==23.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
yarl==1.9.4
